refactor section orden on directories
