+++ sections/home/currentChallenges.section.tsx
