+++ sections/home/ecosystem.section.tsx
