+++ sections/home/mainGoal.section.tsx
