+++ sections/home/partners.section.tsx
