+++ sections/home/whoAreWe.section.tsx
